+++ datashare-index/src/test/java/org/icij/datashare/text/indexing/elasticsearch/DatashareExtractIntegrationTest.java
@@ -50,7 +50,7 @@ public class DatashareExtractIntegrationTest {
         assertThat(doc.getPath()).contains(get("doc.txt"));
         assertThat(doc.getContentEncoding()).isEqualTo(Charset.forName("iso-8859-1"));
         assertThat(doc.getContentType()).isEqualTo("text/plain");
-        assertThat(doc.getExtractionLevel()).isEqualTo(0);
+        assertThat(doc.getExtractionLevel()).isEqualTo((short) 0);
         assertThat(doc.getMetadata()).hasSize(6);
         assertThat(doc.getParentDocument()).isNull();
         assertThat(doc.getRootDocument()).isEqualTo(doc.getId());
